sandbox - move workbench environment service into electron-sandbox
