+++ src/vs/workbench/services/sharedProcess/electron-browser/sharedProcessService.ts
@@ -9,7 +9,8 @@ import { IChannel, IServerChannel, getDelayedChannel } from 'vs/base/parts/ipc/c
 import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/mainProcessService';
 import { ISharedProcessService } from 'vs/platform/ipc/electron-browser/sharedProcessService';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
-import { IWorkbenchEnvironmentService, INativeWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { INativeWorkbenchEnvironmentService } from 'vs/workbench/services/environment/electron-sandbox/environmentService';
 import { IElectronService } from 'vs/platform/electron/electron-sandbox/electron';
 
 export class SharedProcessService implements ISharedProcessService {
