+++ src/vs/workbench/services/timer/electron-browser/timerService.ts
@@ -6,7 +6,8 @@
 import { virtualMachineHint } from 'vs/base/node/id';
 import * as os from 'os';
 import { IElectronService } from 'vs/platform/electron/electron-sandbox/electron';
-import { IWorkbenchEnvironmentService, INativeWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { INativeWorkbenchEnvironmentService } from 'vs/workbench/services/environment/electron-sandbox/environmentService';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IUpdateService } from 'vs/platform/update/common/update';
