+++ src/vs/workbench/workbench.desktop.main.ts
@@ -36,7 +36,7 @@ import 'vs/workbench/electron-browser/desktop.main';
 
 import 'vs/workbench/services/integrity/node/integrityService';
 import 'vs/workbench/services/textMate/electron-browser/textMateService';
-import 'vs/workbench/services/search/node/searchService';
+import 'vs/workbench/services/search/electron-browser/searchService';
 import 'vs/workbench/services/output/electron-browser/outputChannelModelService';
 import 'vs/workbench/services/textfile/electron-browser/nativeTextFileService';
 import 'vs/workbench/services/dialogs/electron-browser/dialogService';
